+++ cpsycle/audio/src/silentdriver.c
@@ -16,7 +16,7 @@ static int driver_open(psy_AudioDriver*);
 static void driver_configure(psy_AudioDriver*, psy_Properties*);
 static int driver_close(psy_AudioDriver*);
 static int driver_dispose(psy_AudioDriver*);
-static unsigned int samplerate(psy_AudioDriver*);
+static uintptr_t samplerate(psy_AudioDriver*);
 static const char* capturename(psy_AudioDriver*, int index);
 static int numcaptures(psy_AudioDriver*);
 static const char* playbackname(psy_AudioDriver*, int index);
@@ -85,7 +85,7 @@ void driver_configure(psy_AudioDriver* driver, psy_Properties* config)
 
 }
 
-unsigned int samplerate(psy_AudioDriver* self)
+uintptr_t samplerate(psy_AudioDriver* self)
 {
 	return 44100;
 }
