+++ cpsycle/driver/mme/mme.c
@@ -138,7 +138,7 @@ static int driver_open(psy_AudioDriver*);
 static int driver_close(psy_AudioDriver*);
 static int driver_dispose(psy_AudioDriver*);
 static void driver_configure(psy_AudioDriver*, psy_Properties*);
-static unsigned int samplerate(psy_AudioDriver*);
+static uintptr_t samplerate(psy_AudioDriver*);
 static const char* capturename(psy_AudioDriver*, int index);
 static int numcaptures(psy_AudioDriver*);
 static const char* playbackname(psy_AudioDriver*, int index);
@@ -351,7 +351,7 @@ void driver_configure(psy_AudioDriver* driver, psy_Properties* config)
 	}
 }
 
-unsigned int samplerate(psy_AudioDriver* self)
+uintptr_t samplerate(psy_AudioDriver* self)
 {
 	return psy_audiodriversettings_samplespersec(&((MmeDriver*)self)->settings);
 }
